这是12月18号晚上七点半的上传

+++ src/main/java/com/qf/pojo/Rider.java
@@ -17,7 +17,7 @@ public class Rider {
     private String riderName;
 
     @Column(name = "id_card")
-    private Integer idCard;
+    private long idCard;
 
     private Integer gender;
